probelms with defineCall error in sequilize, reverting back to old commit
